Done : packages' names optimized
