--- src/com/company/Model/Game/InvalidMoveException.java
@@ -1,7 +1,0 @@
-package com.company.Model.Game;
-
-public class InvalidMoveException extends Exception {
-    InvalidMoveException(String message){
-        super(message);
-    }
-}
