+++ src/org/ta/jcluedo/model/elements/board/Directions.java
@@ -1,4 +1,4 @@
-package com.company.Model.GameBoard;
+package org.ta.jcluedo.model.elements.board;
 
 public enum Directions {
     UP,
