+++ src/org/ta/jcluedo/model/utils/SharedData.java
@@ -1,11 +1,11 @@
-package com.company.Model.Game;
+package org.ta.jcluedo.model.utils;
 
-import com.company.Model.Card.CardDeck;
-import com.company.Model.Card.Place;
-import com.company.Model.Card.Suspect;
-import com.company.Model.Card.Tool;
-import com.company.Model.GameBoard.GameBoard;
-import com.company.Model.Player.Player;
+import org.ta.jcluedo.model.elements.cards.CardDeck;
+import org.ta.jcluedo.model.elements.cards.Place;
+import org.ta.jcluedo.model.elements.cards.Suspect;
+import org.ta.jcluedo.model.elements.cards.Tool;
+import org.ta.jcluedo.model.elements.board.GameBoard;
+import org.ta.jcluedo.model.elements.player.Player;
 
 public class SharedData {
     private static SharedData instance;
