+++ src/org/ta/jcluedo/model/utils/exception/InvalidMoveException.java
@@ -1,0 +1,7 @@
+package org.ta.jcluedo.model.utils.exception;
+
+public class InvalidMoveException extends Exception {
+    public InvalidMoveException(String message){
+        super(message);
+    }
+}
